refactor(cmd): move cmd out of internal package
