+++ cmd/version.go
@@ -11,6 +11,6 @@ var cmdVersion = &cobra.Command{
 	Short: "Print the version number of dargo",
 	Long:  `All software has versions. This is dargo's`,
 	Run: func(cmd *cobra.Command, args []string) {
-		fmt.Println("shadow 0.1.5")
+		fmt.Println("shadow 0.1.6")
 	},
 }
